Fix infinite loop in hashmap lookup
If the hash value was negative, then the value perturb converged
to -1 (instead of 0). But then the perturbation map
  idx -> (5 * idx + perturb + 1)  mod 2^N
degenerated to the linear function
  idx -> (5 * idx )  mod (2^N-1)
which does not have maximal period when 5 divides 2^N-1, so
e.g. for N=4, N=8, ...

The fix consist in making pertub unsigned. And while we are at it,
we also make everything else unsigned for which negative values
make no sense (e.g. idx, capacity, mask, ...)

Fixes #83