Fix requirement checks for HashMap and HashSet
DS_RequiteHashMapOrSet, DS_RequireHashSet and DS_RequireHashMap used to check
the type of the object given as parameter. If the user (or GAP) makes the
object immutable, then the type of the object and then none of the supplied
functions works anymore, because the type of the object changes.

The correct thing to check is the representation filter anyway, because it
reflects the internal representation of the object (in this case IsPosObj).