Disable a test with diffs between GAP 4.10 and 4.11
The disabled test checks whether a positional object that is not a hashmap is
correctly rejected by DS_Hash_Value. However, the error message this prints
differs between GAP 4.10 and 4.11, as the GAP internal name for the type of
component and positional objects was changed.

For now, we work around this by disabling the test. In the future, we might
want to only run it conditionally in some way, or adjust the error message.