Replace constructors PartitionDS with operations
The previous versions of PartitionDS seem to have been written imagining
a future where there would be multiple representations of union-find
data structures. However, at present there's only one, and so it's
annoying to have to include the filter when constructing a PartitionDS.
This does not change any of the existing behaviour, but allows
PartitionDS to be called with a single argument: either the number of
points, or the parts of a partition.